+++ sample/core/navigation/src/main/kotlin/com/mihaicristiancondrea/android/apps/apptoolkit/core/navigation/MainNavigationDefaults.kt
@@ -23,7 +23,7 @@ import androidx.compose.material.icons.outlined.GridView
 import androidx.compose.material.icons.rounded.Apps
 import androidx.compose.material.icons.rounded.GridView
 import com.mihaicristiancondrea.android.apps.apptoolkit.core.ui.R
-import com.mihaicristiancondrea.android.libs.apptoolkit.app.main.domain.models.BottomBarItem
+import com.mihaicristiancondrea.android.libs.apptoolkit.core.ui.models.navigation.BottomBarItem
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.ui.models.navigation.StableNavKey
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf