+++ sample/feature/tiles/src/main/kotlin/com/mihaicristiancondrea/android/apps/apptoolkit/app/tiles/data/repositories/DefaultToolkitTilesRepository.kt
@@ -25,7 +25,7 @@ import com.mihaicristiancondrea.android.apps.apptoolkit.app.tiles.domain.models.
 import com.mihaicristiancondrea.android.apps.apptoolkit.app.tiles.domain.models.ToolkitTileIcon
 import com.mihaicristiancondrea.android.apps.apptoolkit.app.tiles.domain.models.ToolkitTileStatus
 import com.mihaicristiancondrea.android.apps.apptoolkit.app.tiles.domain.models.ToolkitToolKind
-import com.mihaicristiancondrea.android.apps.apptoolkit.app.tiles.domain.models.getTileServiceRequests
+import com.mihaicristiancondrea.android.apps.apptoolkit.app.tiles.service.getTileServiceRequests
 import com.mihaicristiancondrea.android.apps.apptoolkit.core.ui.R
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf